Decompose CmdThrow.py into CmdExplosives, explosion_utils, and slimmed CmdThrow
